qsap: Resolve some compiler warnings
system/qcom/softap/sdk/qsap_api.c:351:59: error: format specifies type 'long' but the argument has type 's32' (aka 'int') [-Werror,-Wformat]
    ALOGD("cmd=%s, Val:%s, INI:%ld \n", pcmd->name, pVal, inifile);
                               ~~~                        ^~~~~~~
                               %d
system/qcom/softap/sdk/qsap_api.c:1308:58: error: format specifies type 'long' but the argument has type 's32' (aka 'int') [-Werror,-Wformat]
    ALOGD("Sap auto channel selection pautochan=%ld \n", *pautochan);
                                                ~~~      ^~~~~~~~~~
                                                %d

+++ softap/sdk/qsap.c
@@ -418,7 +418,7 @@ s32 wifi_qsap_stop_bss(void)
     close(sock);
 
     if (ret) {
-        ALOGE("IOCTL stopbss failed: %d", (int)ret);
+        ALOGE("IOCTL stopbss failed: %d", ret);
         ret = eERR_STOP_BSS;
     } else {
         ALOGD("STOP BSS ISSUED");
